requirements.txt
PyYAML>=6.0
pydantic>=2.0
click>=8.0
pytest>=7.0
pytest-cov>=4.0
black>=23.0
pylint>=2.0
mypy>=1.0
